Upload compiled shaders as artifacts, add asset downloader script
